#225 SponsorsUiState에 대한 PreviewParameterProvider 파일 분리
